Refactor the existing logic
Split out the business logic into a new IntentSender class.

Split out the MethodCall parsing into MethodCallHandlerImpl.
