+++ packages/android_intent/android/build.gradle
@@ -45,3 +45,7 @@ android {
         disable 'InvalidPackage'
     }
 }
+
+dependencies {
+    compileOnly 'androidx.annotation:annotation:1.0.0'
+}
