+++ packages/android_intent/android/src/main/java/io/flutter/plugins/androidintent/AndroidIntentPlugin.java
@@ -4,158 +4,18 @@
 
 package io.flutter.plugins.androidintent;
 
-import android.content.ComponentName;
-import android.content.Context;
-import android.content.Intent;
-import android.net.Uri;
-import android.os.Bundle;
-import android.provider.Settings;
-import android.util.Log;
-import io.flutter.plugin.common.MethodCall;
+import dev.flutter.plugins.androidintent.IntentSender;
+import dev.flutter.plugins.androidintent.MethodCallHandlerImpl;
 import io.flutter.plugin.common.MethodChannel;
-import io.flutter.plugin.common.MethodChannel.MethodCallHandler;
-import io.flutter.plugin.common.MethodChannel.Result;
 import io.flutter.plugin.common.PluginRegistry.Registrar;
-import java.util.ArrayList;
-import java.util.Map;
 
 /** AndroidIntentPlugin */
-@SuppressWarnings("unchecked")
-public class AndroidIntentPlugin implements MethodCallHandler {
-  private static final String TAG = AndroidIntentPlugin.class.getCanonicalName();
-  private final Registrar mRegistrar;
-
+public class AndroidIntentPlugin {
   /** Plugin registration. */
   public static void registerWith(Registrar registrar) {
     final MethodChannel channel =
         new MethodChannel(registrar.messenger(), "plugins.flutter.io/android_intent");
-    channel.setMethodCallHandler(new AndroidIntentPlugin(registrar));
-  }
-
-  private AndroidIntentPlugin(Registrar registrar) {
-    this.mRegistrar = registrar;
-  }
-
-  private String convertAction(String action) {
-    switch (action) {
-      case "action_view":
-        return Intent.ACTION_VIEW;
-      case "action_voice":
-        return Intent.ACTION_VOICE_COMMAND;
-      case "settings":
-        return Settings.ACTION_SETTINGS;
-      case "action_location_source_settings":
-        return Settings.ACTION_LOCATION_SOURCE_SETTINGS;
-      case "action_application_details_settings":
-        return Settings.ACTION_APPLICATION_DETAILS_SETTINGS;
-      default:
-        return action;
-    }
-  }
-
-  private Bundle convertArguments(Map<String, ?> arguments) {
-    Bundle bundle = new Bundle();
-    for (String key : arguments.keySet()) {
-      Object value = arguments.get(key);
-      if (value instanceof Integer) {
-        bundle.putInt(key, (Integer) value);
-      } else if (value instanceof String) {
-        bundle.putString(key, (String) value);
-      } else if (value instanceof Boolean) {
-        bundle.putBoolean(key, (Boolean) value);
-      } else if (value instanceof Double) {
-        bundle.putDouble(key, (Double) value);
-      } else if (value instanceof Long) {
-        bundle.putLong(key, (Long) value);
-      } else if (value instanceof byte[]) {
-        bundle.putByteArray(key, (byte[]) value);
-      } else if (value instanceof int[]) {
-        bundle.putIntArray(key, (int[]) value);
-      } else if (value instanceof long[]) {
-        bundle.putLongArray(key, (long[]) value);
-      } else if (value instanceof double[]) {
-        bundle.putDoubleArray(key, (double[]) value);
-      } else if (isTypedArrayList(value, Integer.class)) {
-        bundle.putIntegerArrayList(key, (ArrayList<Integer>) value);
-      } else if (isTypedArrayList(value, String.class)) {
-        bundle.putStringArrayList(key, (ArrayList<String>) value);
-      } else if (isStringKeyedMap(value)) {
-        bundle.putBundle(key, convertArguments((Map<String, ?>) value));
-      } else {
-        throw new UnsupportedOperationException("Unsupported type " + value);
-      }
-    }
-    return bundle;
-  }
-
-  private boolean isTypedArrayList(Object value, Class<?> type) {
-    if (!(value instanceof ArrayList)) {
-      return false;
-    }
-    ArrayList list = (ArrayList) value;
-    for (Object o : list) {
-      if (!(o == null || type.isInstance(o))) {
-        return false;
-      }
-    }
-    return true;
-  }
-
-  private boolean isStringKeyedMap(Object value) {
-    if (!(value instanceof Map)) {
-      return false;
-    }
-    Map map = (Map) value;
-    for (Object key : map.keySet()) {
-      if (!(key == null || key instanceof String)) {
-        return false;
-      }
-    }
-    return true;
-  }
-
-  private Context getActiveContext() {
-    return (mRegistrar.activity() != null) ? mRegistrar.activity() : mRegistrar.context();
-  }
-
-  @Override
-  public void onMethodCall(MethodCall call, Result result) {
-    Context context = getActiveContext();
-    String action = convertAction((String) call.argument("action"));
-
-    // Build intent
-    Intent intent = new Intent(action);
-    if (mRegistrar.activity() == null) {
-      intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
-    }
-    if (call.argument("flag") != null) {
-      intent.addFlags((Integer) call.argument("flags"));
-    }
-    if (call.argument("category") != null) {
-      intent.addCategory((String) call.argument("category"));
-    }
-    if (call.argument("data") != null) {
-      intent.setData(Uri.parse((String) call.argument("data")));
-    }
-    if (call.argument("arguments") != null) {
-      intent.putExtras(convertArguments((Map) call.argument("arguments")));
-    }
-    if (call.argument("package") != null) {
-      String packageName = (String) call.argument("package");
-      intent.setPackage(packageName);
-      if (call.argument("componentName") != null) {
-        intent.setComponent(
-            new ComponentName(packageName, (String) call.argument("componentName")));
-      }
-      if (intent.resolveActivity(context.getPackageManager()) == null) {
-        Log.i(TAG, "Cannot resolve explicit intent - ignoring package");
-        intent.setPackage(null);
-      }
-    }
-
-    Log.i(TAG, "Sending intent " + intent);
-    context.startActivity(intent);
-
-    result.success(null);
+    IntentSender sender = new IntentSender(registrar.activity(), registrar.context());
+    channel.setMethodCallHandler(new MethodCallHandlerImpl(sender));
   }
 }
